Super refactored...
