+++ KnobScripter/utils.py
@@ -138,3 +138,13 @@ def relistAllKnobScripterPanes():
         if ks.isPane:
             if topParent(ks).metaObject().className() != "Foundry::UI::DockMainWindow":
                 nuke.AllKnobScripters.remove(ks)
+
+
+def clear_layout(layout):
+    if layout is not None:
+        while layout.count():
+            child = layout.takeAt(0)
+            if child.widget() is not None:
+                child.widget().deleteLater()
+            elif child.layout() is not None:
+                clearLayout(child.layout())
